refactor: improve the components placement
