+++ src/constants/colors.ts
@@ -1,0 +1,8 @@
+import colors from 'tailwindcss/colors';
+
+export default {
+  primary: '#036c39',
+  secondary: '#2fc94e',
+  background: '#74b882',
+  ...colors,
+};
